+++ .jenkins/generate_circuit_tests.sh
@@ -18,7 +18,9 @@ salloc -Aproj16 -p\$part -Cnvme -N1 --exclusive --mem=0 \\
                  --output-dir=\$PWD \\
                  --checkpoint-dir=\$PWD \\
                  -p spark.master=spark://\\\$\\(hostname\\):7077 \\
-                 \$RECIPE \$CIRCUIT \$MORPHOS \$TOUCHES
+                 --from \$CIRCUIT All --to \$CIRCUIT All \\
+                 \$RECIPE \$MORPHOS \\
+                 --touches \$TOUCHES
 
 parquet-coalesce circuit.parquet single.parquet
 parquet-compare \\