+++ tests/conftest.py
@@ -17,7 +17,9 @@ DATADIR = os.path.join(os.path.dirname(__file__), "circuit_1000n")
 ARGS = (
     os.path.join(DATADIR, "builderRecipeAllPathways.xml"),
     (os.path.join(DATADIR, "nodes.h5"), "All"),
+    (None, None),
     (os.path.join(DATADIR, "nodes.h5"), "All"),
+    (None, None),
     os.path.join(DATADIR, "morphologies/h5"),
     os.path.join(DATADIR, "touches/*.parquet")
 )