+++ tests/test_morpho_shift.py
@@ -11,6 +11,7 @@ from recipe import Recipe
 import sparkmanager as sm
 
 from spykfunc.circuit import Circuit
+from spykfunc.utils.conf import Configuration
 
 
 class MockLoader:
@@ -35,7 +36,9 @@ def test_shift():
     from spykfunc.filters.implementations.synapse_reposition import SynapseReposition
     from spykfunc.dataio.morphologies import MorphologyDB
 
-    sm.create("test_shift")
+    conf = Configuration(".")
+
+    sm.create("test_shift", conf("spark"))
 
     neurons = sm.read.json(sm.parallelize(NEURONS))
     touches = sm.read.json(sm.parallelize(TOUCHES))
@@ -53,29 +56,24 @@ def test_shift():
         Path(__file__).parent / "circuit_O1_partial" / "morphologies" / "h5"
     )
 
-    touches.show()
     fltr = SynapseReposition(recipe, c.source, c.target, c.morphologies)
-    result = fltr.apply(c)
-    result.select(touches.columns).show()
+    result = fltr.apply(c).select(touches.columns)
 
-    shifted = result.where(result.src == 39167)
-    assert shifted.select("post_section").distinct().rdd.keys().collect() == [1]
-    assert shifted.select("post_segment").distinct().rdd.keys().collect() == [0]
-    (offset,) = shifted.select("post_offset").distinct().rdd.keys().collect()
+    shifted = result.where(result.src == 39167).toPandas()
+    assert shifted["post_section"].unique() == [1]
+    assert shifted["post_segment"].unique() == [0]
+    (offset,) = shifted["post_offset"].unique()
     assert abs(offset - 0.5) < 1e-5
-    (fraction,) = shifted.select("post_section_fraction").distinct().rdd.keys().collect()
+    (fraction,) = shifted["post_section_fraction"].unique()
     assert abs(fraction - 0.00353) < 1e-5
-    (dist,) = shifted.select("distance_soma").distinct().rdd.keys().collect()
+    (dist,) = shifted["distance_soma"].unique()
     assert abs(dist - 0.5) < 1e-5
 
-    untouched = result.where(result.src == 101)
-    assert untouched.select("post_section").distinct().rdd.keys().collect() == [337]
-    assert untouched.select("post_segment").distinct().rdd.keys().collect() == [4]
-    assert untouched.select("post_offset").distinct().rdd.keys().collect() == [0.012562983]
-    (fraction,) = untouched.select("post_section_fraction").distinct().rdd.keys().collect()
-    assert abs(fraction - 0.666) < 1e-5
-    (dist,) = untouched.select("distance_soma").distinct().rdd.keys().collect()
-    assert abs(dist) > 1
+    untouched_before = touches.where(touches.src == 101).toPandas()
+    untouched_after = result.where(result.src == 101).toPandas()
+
+    print(untouched_after.compare(untouched_before))
+    assert untouched_after.equals(untouched_before)
 
 
 RECIPE = """
@@ -149,6 +147,7 @@ TOUCHES = [
             "dst": 42113,
             "pre_section": 8,
             "pre_segment": 2,
+            "post_branch_type": 3,
             "post_section": 337,
             "post_section_fraction": 0.666,
             "post_segment": 4,